Made some changes, added oscilloscope reset button.

+++ UnifiedTestSuiteApp/MainWindow_GlobalObjects.cs
@@ -15,9 +15,9 @@ namespace UnifiedTestSuiteApp
     public partial class MainWindow : Window
     {
         private const string appName = "S.A.L.T Application";
-        private const int refreshIntervalOneTwoChannels = 200;  // the graph refresh interval in ms for when 1-2 channels are enabled
-        private const int refreshIntervalThreePlusChannels = 300;
-        private int refreshInterval = refreshIntervalOneTwoChannels;  // just to start. Please don't change this manually without a reason
+        private const double refreshIntervalOneTwoChannels = 200;  // the graph refresh interval in ms for when 1-2 channels are enabled
+        private const double refreshIntervalThreePlusChannels = 300;
+        private double refreshInterval = refreshIntervalOneTwoChannels;  // just to start. Please don't change this manually without a reason
         // updates when we have 3-4 channels enabled)
         private static System.Timers.Timer refreshTimer;    // gotta make sure there's no ambiguity with the threading timer
         private readonly IOscilloscope scope;               //
